expansion v3.5
